backend: working GET and POST!

+++ backend/src/core/users.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, Debug, sqlx::Type)]
 #[sqlx(transparent)]
-pub struct UserId(pub i64);
+pub struct UserId(pub i32);
 
 #[derive(Deserialize, Serialize, Debug)]
 pub struct User {
